# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SATPAM API",
    "description": "Search-based AI Threat Prevention and Mapping — Prototype backend untuk sistem graph intelligence deteksi risiko judol/pinjol ilegal. Semua data bersifat dummy/simulasi.",
    "version": "0.1.0-week1"
  },
  "paths": {
...
    "/api/verification-cases": {
      "get": {
        "tags": [
          "verification"
        ],
        "summary": "List Verification Cases",
        "description": "Daftar verification case dengan filter status/risk level.",
        "operationId": "list_verification_cases_api_verification_cases_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/verification-cases/{case_id}": {
      "get": {
        "tags": [
          "verification"
        ],
        "summary": "Get Verification Case",
        "description": "Detail satu verification case.",
        "operationId": "get_verification_case_api_verification_cases__case_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "verification"
        ],
        "summary": "Update Verification Case",
        "description": "Update status verifikasi dan catatan reviewer; dicatat ke audit log.",
        "operationId": "update_verification_case_api_verification_cases__case_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/traffic-events": {
      "get": {
        "tags": [
          "intel"
        ],
        "summary": "List Traffic Events",
        "description": "Daftar traffic event simulasi.",
        "operationId": "list_traffic_events_api_traffic_events_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/traffic-events/import": {
      "post": {
        "tags": [
          "intel"
        ],
        "summary": "Import Traffic Events",
        "description": "Import traffic log simulasi (admin). simulationOnly wajib True.",
        "operationId": "import_traffic_events_api_traffic_events_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrafficImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/crawler-findings": {
      "get": {
        "tags": [
          "intel"
        ],
        "summary": "List Crawler Findings",
        "description": "Daftar crawler finding dummy.",
        "operationId": "list_crawler_findings_api_crawler_findings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "finding_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Finding Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/crawler-findings/import": {
      "post": {
        "tags": [
          "intel"
        ],
        "summary": "Import Crawler Findings",
        "description": "Import crawler finding dummy (admin). simulationOnly wajib True.",
        "operationId": "import_crawler_findings_api_crawler_findings_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlerImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfidenceLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "ConfidenceLevel"
      },
      "CrawlerFindingNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "const": "CrawlerFinding",
            "title": "Type",
            "default": "CrawlerFinding"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "firstSeenAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstseenat"
          },
          "lastSeenAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastseenat"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "riskScore": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Riskscore"
          },
          "riskLevel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "verificationStatus": {
            "$ref": "#/components/schemas/VerificationStatus",
            "default": "unreviewed"
          },
          "findingType": {
            "type": "string",
            "title": "Findingtype"
          },
          "sourceUrl": {
            "type": "string",
            "title": "Sourceurl"
          },
          "contentSummary": {
            "type": "string",
            "title": "Contentsummary"
          },
          "matchedKeywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matchedkeywords",
            "default": []
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Capturedat"
          },
          "simulationOnly": {
            "type": "boolean",
            "const": true,
            "title": "Simulationonly",
            "default": true
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "label",
          "source",
          "createdAt",
          "findingType",
          "sourceUrl",
          "contentSummary",
          "capturedAt"
        ],
        "title": "CrawlerFindingNode",
        "description": "Temuan crawler dummy — simulation_only wajib True."
      },
      "CrawlerImportRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CrawlerFindingNode"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CrawlerImportRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel"
      },
...
      "TrafficEventNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "const": "TrafficEvent",
            "title": "Type",
            "default": "TrafficEvent"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "firstSeenAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstseenat"
          },
          "lastSeenAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastseenat"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "riskScore": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Riskscore"
          },
          "riskLevel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "verificationStatus": {
            "$ref": "#/components/schemas/VerificationStatus",
            "default": "unreviewed"
          },
          "eventType": {
            "type": "string",
            "title": "Eventtype"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "sourceAlias": {
            "type": "string",
            "title": "Sourcealias"
          },
          "destinationDomain": {
            "type": "string",
            "title": "Destinationdomain"
          },
          "requestCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Requestcount",
            "default": 0
          },
          "simulationOnly": {
            "type": "boolean",
            "const": true,
            "title": "Simulationonly",
            "default": true
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "label",
          "source",
          "createdAt",
          "eventType",
          "timestamp",
          "sourceAlias",
          "destinationDomain"
        ],
        "title": "TrafficEventNode",
        "description": "Log trafik simulasi — simulation_only wajib True."
      },
      "TrafficImportRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrafficEventNode"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TrafficImportRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "unreviewed",
          "needs_review",
          "verified_risk",
          "false_positive",
          "escalated",
          "closed"
        ],
        "title": "VerificationStatus"
      },
      "VerificationUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "decisionNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisionnote"
          },
          "reviewerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewerid"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "VerificationUpdate"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}